fix(runsh): refactored for optimization
Refactored the run.sh script so that it will ask user for a path to the
python version needed when installing the virtual environment. It also
has better error handling. It installs only the versions specified in
the requirements file for python packages rather than searching for
updates or upgrades.